Moved some test util method to auth util
